Fixed several issues with new feature vector space
Main problem was due to the use of int rather than double for FEATURE_DEPTH, so band calculations didnt work

+++ featurevector.h
@@ -31,7 +31,7 @@ public:
     static unsigned int NUMBER_OF_FEATURES;
     static unsigned int NUMBER_OF_FEATURE_VECTORS;
     static double       FEATURE_RANGE;
-    static unsigned int FEATURE_DEPTH;                      /***bit depth of features**/
+    static double       FEATURE_DEPTH;                      /***bit depth of features**/
 
     virtual unsigned int GetValue() const;
     virtual unsigned int GetLength() const;
